# docs/swagger.yaml
basePath: /
definitions:
...
  dto.BackupResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  dto.DashboardStatsResponse:
    properties:
      devices_with_installation:
        type: integer
      devices_without_installation:
        type: integer
      suspended_users:
        type: integer
      total_devices:
        type: integer
      total_groups:
        type: integer
      total_users:
        type: integer
      total_vehicles:
        type: integer
      users_with_installations:
        type: integer
    type: object
...
  dto.GroupAssignDeviceRequest:
    properties:
      device_imei:
        type: string
      group_id:
        type: string
    required:
    - device_imei
    - group_id
    type: object
  dto.GroupRemoveDeviceRequest:
    properties:
      device_imei:
        type: string
      group_id:
        type: string
    required:
    - device_imei
    - group_id
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8888
info:
  contact:
    email: mddx56@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestionar motos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Motos API
  version: 0.1.2
paths:
...
  /api/backup/download:
    get:
      description: Descarga el archivo de respaldo completo (.sql) más reciente desde
        el servidor externo.
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Archivo SQL de respaldo
          schema:
            type: file
        "401":
          description: No autorizado
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Acceso denegado (Solo Admin)
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Error interno o al descargar
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Descargar Backup Completo
      tags:
      - backup
  /api/backup/schema/download:
    get:
      description: Descarga el archivo de respaldo del esquema (.sql) más reciente
        desde el servidor externo.
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Archivo SQL de esquema
          schema:
            type: file
        "401":
          description: No autorizado
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Acceso denegado (Solo Admin)
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Error interno o al descargar
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Descargar Backup del Esquema
      tags:
      - backup
  /api/backup/schema/trigger:
    post:
      consumes:
      - application/json
      description: Inicia el proceso de respaldo *solo del esquema* (sin datos) en
        el servidor externo.
      produces:
      - application/json
      responses:
        "200":
          description: Proceso iniciado correctamente
          schema:
            $ref: '#/definitions/dto.BackupResponse'
        "401":
          description: No autorizado
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Acceso denegado (Solo Admin)
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Iniciar Backup del Esquema
      tags:
      - backup
  /api/backup/trigger:
    post:
      consumes:
      - application/json
      description: Inicia el proceso de respaldo completo de la base de datos en el
        servidor externo.
      produces:
      - application/json
      responses:
        "200":
          description: Proceso iniciado correctamente
          schema:
            $ref: '#/definitions/dto.BackupResponse'
        "401":
          description: No autorizado
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Acceso denegado (Solo Admin)
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Iniciar Backup Completo
      tags:
      - backup
  /api/dashboard/stats:
    get:
      description: Retorna estadísticas básicas del sistema para el dashboard administrativo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DashboardStatsResponse'
              type: object
        "401":
          description: No autorizado
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Acceso denegado (Solo Admin)
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Obtener Estadísticas del Dashboard
      tags:
      - dashboard
...
  /api/group:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all groups
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: User ID Filter
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all groups
      tags:
      - groups
  /api/group/assign:
    post:
      consumes:
      - application/json
      description: Assign a device to a specific group
      parameters:
      - description: Assignment Request
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/dto.GroupAssignDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Assign a device to a group
      tags:
      - groups
  /api/group/remove:
    post:
      consumes:
      - application/json
      description: Remove a device from a specific group
      parameters:
      - description: Remove Request
        in: body
        name: removel
        required: true
        schema:
          $ref: '#/definitions/dto.GroupRemoveDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Remove a device from a group
      tags:
      - groups
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"